requirements.txt
streamlit
Pillow
streamlit_extras